// types.ts
export interface SubMenuItem {
  label: string;
  icon?: ReactNode;
  shortcut?: string[];
  onClick?: () => void;
  danger?: boolean;
}

export interface MenuItemDef {
  label: string;
  icon?: ReactNode;
  subItems?: SubMenuItem[];
  shortcut?: string[];
  onClick?: () => void;
  danger?: boolean;
}